+++ oak_functions/experimental/tf_inference/Cargo.toml
@@ -8,10 +8,14 @@ license = "Apache-2.0"
 [dependencies]
 anyhow = "*"
 bytes = "*"
+bincode = "*"
 log = "*"
 oak_functions_abi = { path = "../../abi" }
+oak_functions_extension = { path = "../../extension" }
 oak_logger = { path = "../../logger" }
 prost = "*"
 serde = "*"
 serde_derive = "*"
 tract-tensorflow = "*"
+# Use wasmi in `no_std` mode.
+wasmi = { version = "*", default-features = false, features = ["core"] }
